Git is one control with two faces
Hickey fragmentation / Lowy parallel receptacle: GitNews was a second
createCommit, a second popover, and the same panel. The desktop pill and
the phone banner are two faces of Commit. The crush-era saysPhone path
goes with it — the pill no longer sits on a phone.
